+++ crates/solver/src/driver.rs
@@ -86,7 +86,7 @@ impl Driver {
         max_settlement_price_deviation: Option<Ratio<BigInt>>,
         token_list_restriction_for_price_checks: PriceCheckTokens,
         tenderly: Option<Arc<dyn TenderlyApi>>,
-        solution_comparison_decimal_cutoff: i32,
+        solution_comparison_decimal_cutoff: u16,
     ) -> Self {
         let post_processing_pipeline = PostProcessingPipeline::new(
             native_token,
